Create Code Challenge 1 Steven's Tip Calculator
My name is Scott Fennell (U86364709) and I am student at USF in the ISM 3232-521 course. This coding challenge is called the "Steven's Tip Calculator Challenge" This project will be able to calculator and give output of the bill amount, tip amount, and total amount of the bill. The calculations will be based on numerical ranges to determine what percentage to be used for the tip.